refactor: rangement des features par sections de navigation
