+++ src/features/Settings/About/AboutScreen.tsx
@@ -4,11 +4,11 @@ import { SafeAreaView, SafeAreaInsetsContext } from 'react-native-safe-area-cont
 import Constants from 'expo-constants';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 
-import style, { tokens } from '../../shared/theme/Theme';
-import Translator from '../../shared/i18n/Translator';
-import { AppContext } from '../../shared/services/AppCore';
-import { URL } from '../../shared/services/DataService';
-import { withHeaderAnimation } from '../../shared/navigation/NavHelpers';
+import style, { tokens } from '../../../shared/theme/Theme';
+import Translator from '../../../shared/i18n/Translator';
+import { AppContext } from '../../../shared/services/AppCore';
+import { URL } from '../../../shared/services/DataService';
+import { withHeaderAnimation } from '../../../shared/navigation/NavHelpers';
 
 export interface SectionProps {
     title: string;
@@ -131,7 +131,7 @@ class AboutScreen extends React.Component<AboutScreenProps> {
                             justifyContent: 'center', alignItems: 'center'
 						}}>
                             <Image
-                                source={require('../../../assets/icons/logo.png')}
+                                source={require('../../../../assets/icons/logo.png')}
                                 style={{ width: 70, height: 70, resizeMode: 'contain' }}
                             />
 						</View>
