+++ src/shared/navigation/StackNavigator.tsx
@@ -7,7 +7,7 @@ import { RouteProp } from '@react-navigation/native';
 import MainTabNavigator from './MainTabNavigator';
 import GroupSearch from '../../features/Planning/screens/GroupSelectionScreen';
 import Group from '../../features/Planning/screens/ScheduleScreen';
-import About from '../../features/About/AboutScreen';
+import About from '../../features/Settings/About/AboutScreen';
 import Settings from '../../features/Settings/SettingsScreen';
 import CredentialsSettingsScreen from '../../features/Scolarite/CredentialsSettingsScreen';
 import { CredentialsProvider } from '../../features/Scolarite/services/CredentialsContext';
@@ -15,14 +15,14 @@ import WebBrowser from '../../features/Browser/WebBrowserScreen';
 import Geolocation from '../../features/Map/MapScreen';
 import Course from '../../features/Planning/components/CourseCard';
 import DayView from '../../features/Planning/views/DayView';
-import CrousScreen from '../../features/Crous/CrousScreen';
-import CrousMenuScreen from '../../features/Crous/CrousMenuScreen';
-import LibraryScreen from '../../features/Library/LibraryScreen';
-import LibraryDetailsScreen from '../../features/Library/LibraryDetailsScreen';
-import BdeDetailsScreen from '../../features/Bde/BdeDetailsScreen';
-import BdeScreen from '../../features/Bde/BdeScreen';
-import FreeRoomScreen from '../../features/FreeRoom/FreeRoomScreen';
-import FreeRoomDetailsScreen from '../../features/FreeRoom/FreeRoomDetailsScreen';
+import CrousScreen from '../../features/Campus/Crous/CrousScreen';
+import CrousMenuScreen from '../../features/Campus/Crous/CrousMenuScreen';
+import LibraryScreen from '../../features/Campus/Library/LibraryScreen';
+import LibraryDetailsScreen from '../../features/Campus/Library/LibraryDetailsScreen';
+import BdeDetailsScreen from '../../features/Campus/Bde/BdeDetailsScreen';
+import BdeScreen from '../../features/Campus/Bde/BdeScreen';
+import FreeRoomScreen from '../../features/Campus/FreeRoom/FreeRoomScreen';
+import FreeRoomDetailsScreen from '../../features/Campus/FreeRoom/FreeRoomDetailsScreen';
 
 import style, { tokens } from '../theme/Theme';
 import { AppContext, treatTitle } from '../services/AppCore';
